Move MakeOrderRequest to requests' DTO and update imports
